requirements.txt
dash
dash_ag_grid
dash-bootstrap-components
dash_daq
selenium
astor
Pillow
openai
numpy
pandas
tqdm